+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2021-07-15
 - Copying depth buffer from G-buffer to default buffer so that we can forward render
   with the correct depths after deferred rendering.
+- Skybox implementation
 
 ### 2021-07-11
 - Spotlights for deferred renderer.